requirements.txt
# Code Repository Semantic Search
# Session 7 Assignment - RAG Memory, Context, and State Management

# Core dependencies
anthropic>=0.45.0
mcp>=1.0.0
python-dotenv>=1.0.0

# Vector search and embeddings
faiss-cpu>=1.7.4
numpy>=1.24.0

# Data validation
pydantic>=2.0.0

# AST parsing - Core (required)
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0

# AST parsing - Optional languages (install as needed)
# Uncomment to enable AST support for these languages:
# tree-sitter-javascript>=0.23.0
# tree-sitter-typescript>=0.23.0
# tree-sitter-java>=0.23.0
# tree-sitter-rust>=0.23.0
# tree-sitter-go>=0.23.0
# tree-sitter-c>=0.23.0
# tree-sitter-cpp>=0.23.0
# tree-sitter-ruby>=0.23.0

# Git operations
gitpython>=3.1.0

# Terminal UI
rich>=13.0.0

# HTTP requests (for Ollama embeddings)
requests>=2.31.0
